Totally refactored a project to be able to build a package. Added setup.cfg and pyproject.toml. Structured requirements.txt. Updated copyright notes in accordance with Apache License recommended way.
