--- core/__init__.py
@@ -1,7 +1,0 @@
-
-#  Copyright (c) 2022. Eugene Popov
-#  All rights reserved.
-
-from core.batch import Batch
-
-
