--- extractors/file.py
@@ -1,16 +1,0 @@
-
-#  Copyright (c) 2022. Eugene Popov
-#  All rights reserved.
-
-from extractors.basic import FileExtractor
-
-
-class ExcelFileExtractor(FileExtractor):
-
-    def _set_extraction_timestamp(self):
-        pass
-
-    def extract(self):
-        pass
-
-
