+++ pyproject.toml
@@ -1,0 +1,6 @@
+[build-system]
+requires = [
+    "setuptools>=62",
+    "wheel"
+]
+build-backend = "setuptools.build_meta"
