+++ requirements.txt
@@ -1,12 +1,17 @@
 
-numpy==1.22.3; python_version >= '3.9'
-pandas==1.3.5
-SQLAlchemy==1.3.24
-psycopg2-binary==2.8.6
-pytz==2022.1
+# essential
 requests==2.27.1
-pyodbc==4.0.32
 pytest==7.1.2
 pyyaml==6.0
 pyarrow==8.0.0
+
+# additional
+numpy==1.22.3; python_version >= '3.9'
+pandas==1.3.5
+psycopg2-binary==2.8.6
+SQLAlchemy==1.3.24
+pyodbc==4.0.32
 openpyxl==3.0.9
+
+# deprecated
+pytz==2022.1
