+++ setup.cfg
@@ -1,0 +1,48 @@
+[metadata]
+name = witness-etl
+version = 0.0.1
+author = Eugene Popov
+author_email = evgeniypalych@gmail.com
+description = A minimal ETL library.
+long_description = file: readme.md
+long_description_content_type = text/markdown
+license = Apache License 2.0
+license_files =
+   LICENSE
+classifiers =
+    Intended Audience :: Developers
+    Programming Language :: Python :: 3.9
+    License :: OSI Approved :: Apache Software License
+    Operating System :: OS Independent
+project_urls =
+    Source Code=https://github.com/eppv/witness
+
+[options]
+packages = find:
+python_requires = >=3.9
+include_package_data = True
+setup_requires =
+    wheel
+install_requires =
+    pytest==7.1.2
+    pyyaml>=6.0
+    pyarrow>=8.0.0
+    pyodbc==4.0.32
+    numpy==1.22.3; python_version >= "3.9"
+    pandas>=1.3.5, <1.4
+    sqlalchemy>=1.3.24, <1.4
+    psycopg2-binary==2.8.6
+    requests==2.27.1
+
+
+[options.packages.find]
+include =
+    witness*
+exclude =
+    temp*
+    scripts*
+    docs*
+    witness.tests*
+
+[files]
+packages = witness
