--- tests/loaders/__init__.py
@@ -1,3 +1,0 @@
-#  Copyright (c) 2022. Eugene Popov
-#  All rights reserved.
-
