+++ witness/tests/loaders/pandas.py
@@ -1,10 +1,21 @@
 
-#  Copyright (c) 2022. Eugene Popov
-#  All rights reserved.
+#  Copyright (c) 2022.  Eugene Popov.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#     you may not use this file except in compliance with the License.
+#     You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#     Unless required by applicable law or agreed to in writing, software
+#     distributed under the License is distributed on an "AS IS" BASIS,
+#     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#     See the License for the specific language governing permissions and
+#     limitations under the License.
 
 import pytest
-from loaders.pandas import PandasSQLLoader
-from tests.core.batch import mock_batch
+from witness.loaders import PandasSQLLoader
+from core.batch import mock_batch
 from sqlalchemy import create_engine
 
 import yaml
